+++ system/src/Grav/Common/Grav.php
@@ -55,6 +55,7 @@ class Grav extends Container
      *            to the dependency injection container.
      */
     protected static $diMap = [
+        'Grav\Common\Service\AccountsServiceProvider',
         'Grav\Common\Service\AssetsServiceProvider',
         'Grav\Common\Service\BackupsServiceProvider',
         'Grav\Common\Service\ConfigServiceProvider',
@@ -68,7 +69,6 @@ class Grav extends Container
         'Grav\Common\Service\SetupServiceProvider',
         'Grav\Common\Service\StreamsServiceProvider',
         'Grav\Common\Service\TaskServiceProvider',
-        'Grav\Common\Service\UserServiceProvider',
         'browser'    => 'Grav\Common\Browser',
         'cache'      => 'Grav\Common\Cache',
         'events'     => 'RocketTheme\Toolbox\Event\EventDispatcher',