+++ system/src/Grav/Common/User/FlexUser/UserIndex.php
@@ -113,7 +113,7 @@ class UserIndex extends FlexIndex
         // Load saved index file.
         $grav = Grav::instance();
         $locator = $grav['locator'];
-        $filename = $locator->findResource('user-data://flex/indexes/users.yaml', true, true);
+        $filename = $locator->findResource('user-data://flex/indexes/accounts.yaml', true, true);
 
         return CompiledYamlFile::instance($filename);
     }